Convert facebook to higher-order component
