+++ src/components/theme-legacy/facebook-button.js
@@ -1,6 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
+import { withFacebook } from '../../containers/hoc-facebook'
+
 const FacebookButton = ({ shareFacebook }) => (
   <button
     id='facebook-button'
@@ -15,4 +17,4 @@ FacebookButton.propTypes = {
   shareFacebook: PropTypes.func
 }
 
-export default FacebookButton
+export default withFacebook(FacebookButton)
